[BUGFIX] Follow-up: Add setter for $cObj in SimpleProcessor as well
Since TYPO3 11.4 and https://forge.typo3.org/issues/94956 the use of the
public property $cObj in data processors is deprecated. As a
replacement, a public setter "setContentObjectRenderer" should be used.
This commit adds this setter but keeps the public property as-is to
assure compatibility with TYPO3 < 11.4.

This is a follow-up to 2d1704c.